feat: add vitest tests
