+++ package.json
@@ -8,7 +8,9 @@
     "build": "tsc && vite build",
     "lint": "eslint ./src",
     "lint:fix": "eslint ./src --fix --color",
-    "test": "jest --passWithNoTests",
+    "test": "vitest",
+		"coverage": "vitest run --coverage",
+    "test:jest": "jest --passWithNoTests",
     "preview": "vite preview",
     "prepare": "husky install",
     "pre-commit": "lint-staged"
@@ -37,13 +39,16 @@
     "@babel/preset-react": "^7.18.6",
     "@babel/preset-typescript": "^7.21.0",
     "@jest/globals": "^29.5.0",
+    "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
     "@types/jest": "^29.5.0",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
     "@typescript-eslint/eslint-plugin": "^5.55.0",
     "@typescript-eslint/parser": "^5.55.0",
     "@vitejs/plugin-react": "^3.1.0",
+    "@vitest/coverage-c8": "^0.29.3",
     "babel-jest": "^29.5.0",
     "eslint": "^8.36.0",
     "eslint-config-prettier": "^8.7.0",
@@ -58,11 +63,13 @@
     "husky": "^8.0.3",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
+    "jsdom": "^21.1.1",
     "lint-staged": "^13.2.0",
     "prettier": "^2.8.4",
     "react-test-renderer": "^18.2.0",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.2",
-    "vite": "^4.2.0"
+    "vite": "^4.2.0",
+    "vitest": "^0.29.3"
   }
 }
